# ~10x speedup for selecting traces from Stream by non-wildcarded trace ID