… obspy/core/tests/test_stream.py