### What does this PR do?
Speed up the simple selection of traces in a stream by their fixed trace-ID (i.e., using `st.select(id=...)`, with no wildcards or single NSLC-codes involved). Test on a 512-trace stream shows a >10x speedup.

### Why was it initiated?  Any relevant Issues?
`stream.select`is the most concise and convenient function to select specific traces from a stream, but if that function is called on larger streams (~> 100 traces) it can be surprisingly slow. The reason appears to be that it uses several invocations of `fnmatch.fnmatch` for each trace and for all string comparisons even if there are no wildcards. For this type of usage, it's easy to just revert to standard string-comparison in python. Some codes use the `select` so often that this PR makes a noteable difference for total execution time (e.g., ~35 % speedup for `eqcorrscan.utils.catalog_to_dd.write_correlations`).

Demonstration of the speedup for `st.select()` (test on master-branch, this PR speeds up formulation A to 0.45 ms per loop):
```python
import obspy
st = obspy.read()
for j in range(1,10):
    st = st + st

# A: 4.77 ms per loop:
%timeit st.select(id='BW.RJOB..EHZ')
# B: 2.9 ms per loop:
%timeit obspy.Stream([tr for tr in st  if tr.id=='BW.RJOB..EHZ'])
# C: 0.36 ms per loop:
%timeit obspy.Stream([tr for tr in st  if tr.stats.network=='BW' and tr.stats.station == 'RJOB' and tr.stats.location == '' and tr.stats.channel=='EHZ'])
```

### PR Checklist
- [X] Correct base branch selected? `master` for new features, `maintenance_...` for bug fixes
- [X] This PR is not directly related to an existing issue (which has no PR yet).
- [ ] If the PR is making changes to documentation, docs pages can be built automatically.
      Just remove the space in the following string after the + sign: "+ DOCS"
- [ ] If any network modules should be tested for the PR, add them as a comma separated list
      (e.g. `clients.fdsn,clients.arclink`) after the colon in the following magic string: "+TESTS:"
      (you can also add "ALL" to just simply run all tests across all modules)
- [X] All tests still pass. (Ubuntu tests here on github succeed)
- [X] Any new features or fixed regressions are be covered via new tests.
- [ ] Any new or changed features have are fully documented.
- [ ] Significant changes have been added to `CHANGELOG.txt` .
- [ ] First time contributors have added your name to `CONTRIBUTORS.txt` .