@flixha 2021-11-02
> Any idea what's going on with the macos fails?

I am really unsure what those mean, I'm hoping someone may be able to help. It doesn't seem like those were directly related to any small changes that I did with this PR. I tested this PR on my own Mac (MacOS 10.13.7 / Darwin 17.7.0) for both python versions, and both succeeded:
https://tests.obspy.org/114682/ (python3.8)
https://tests.obspy.org/114684/ (python3.7)

The tests runs that fail produce lots of error messages similar to `TypeError: int() argument must be a string, a bytes-like object or a number, not '_NoValueType'` (or with `float`) in calls to `max()` in numpy. Looking for any differences between the failed runs and my runs, I can see a tiny difference for numpy (GH-Action-runs with numpy1.21.2 failed, my own with numpy1.21.3 worked), and of course there's the difference for the actions running on macOS 10.15.6 beta 2: https://tests.obspy.org/114657/

I can at least find one successful recent test on macOS 11.5 beta2 (Darwin 20.6.0): https://tests.obspy.org/114619/

I just triggered a new Action run to see if there's any change, let's see.

I don't have access to a system with macOS 10.15.6 beta 2 to test (why beta?!), and I don't think I have any access to the Obspy-actions to use these for debugging? Maybe someone has other suggestions how to debug on a specific system?