@d-chambers 2021-11-02
@flixha I made a "hard fork" of ObsPy on my personal github account (hard so the CIs are not connected) where I test the CI. You can do this by cloning ObsPy, then adding another remote pointing to your fork and manually pushing. Then you can test CI on your own and it wont notify all the ObsPy watchers each time you push to your own fork. 

The fails are clearly not related to your PR though.