@d-chambers 2021-11-02
This looks like a good contribution. One question:

How much does this slow down `select` when wildcards are used?