@flixha 2021-11-03
@d-chambers thanks for the point on the forks. I guess I kind of have set up a hard fork, but haven't really used its full potential for the actions, so it's great to get a tip in the right direction. I'll try that out!

Good question regarding using the wildcards, here's the numbers:
Previously: `%timeit st.select(id='B?.RJO?.*.EHZ')` --> 4.0 ms per loop
With this PR:`%timeit st.select(id='B?.RJO?.*.EHZ')` --> 4.2 ms per loop (i.e., 5 % longer runtime)