@d-chambers 2021-11-03
Ok, I think it is worth 5% slower wildcard queries for the 10x speed up without them. Anyone else have an opinion?