@flixha 2021-11-08
> Ok, I guess no one else cares. Can you add a quick changelog entry then I think we are good to merge.

Thanks for inviting to get more feedback and reviewing the PR! I added a line to the Changelog now, so then it should be ready to go.