@ThomasLecocq 2022-01-04
@d-chambers scrolling through the PRs and this is indeed unrelated to MacOS sh**t testing , and I don't see why macos or win would behave differently here (nothing related to paths or stdlib) - good to go for me