@claudiodsf 2022-01-04
Continous Integration works again, thanks to @ThomasLecocq! 🥳 

@flixha: could you please rebase and force push, to trigger all the tests?