@flixha 2022-01-05
@claudiodsf  claudiodsf 
> Continous Integration works again, thanks to ThomasLecocq! 
> 
> @flixha: could you please rebase and force push, to trigger all the tests?

I rebased and triggered the tests to run again.