@flixha 2022-01-05
@ThomasLecocq @claudiodsf
> @claudiodsf why is the changed files list soooo long ? not forced-rebased ?

Yes that was the mistake. I fixed it now.