@ThomasLecocq 2022-01-05
ok, 🟢 light from the CI, merging.

For the future, was wondering if there wouldn't be a more clever way of using `re` for matching trace ids than sequentially over each parameter. Aaaanyway, for later.